Node monitor.
Also made a new directory tordiffs for diffs that
don't/won't/can't/shouldn't/didn't (yet?) make it into the proper Tor
distribution. 

or-conn-bw.diff is a diff to enable bandwidth monitoring for OR
connections. 

nodemon spits out stats about how much bandwidth is going to each or
connection, and also ORCONN failure statistics and circuit death info. On my
slow node the relative bandwidth rates do not match up with the rest of the
driectory, at least not after measurements covering durations on the order of
days.. Maybe this will change when I get a fast node.



git-svn-id: https://svn.torproject.org/svn/torflow/trunk@9469 55e972cd-5a19-0410-ae62-a4d7a52db4cd