HTTP 3xx redirect handling
Added new inconclusive and failure cases for redirects. Non-tor
redirects return an inconclusive test result and an attempt to add the
redirect location to the target list. Previously when the HTTP test
requested a binary file that redirected to a dynamic html page, the
binary file's url would be marked as FailureDynamic. Now it'll be
marked as InconclusiveRedirect. Tor only redirects result in
FailureRedirect.