Workaround strange issue with unmeasured flag parsing.
It seems like consensus/newdesc events are causing TorCtl to reset the
unmeasured flag when we shouldn't be. That is a TorCtl bug, but since there
may also be memory bloat issues with leaving the bwauthority_child processes
around anyway, just exit and let the parent restart the slice with a clean
consensus.